Reorganize forward declarations
